First stages of spike ball + better transitions
Added in new powerup, with stubs for later use, and updated the goal scoring transitions to include the players as well (While disregarding other modifications, to prevent accidental coroutine stacking of darkness modifiers. Special transitions are handed in a "TotalLerp" float inside of the match players script, which global modifies all darkness cutouts.